+++ src/axom/core/ArrayBase.hpp
@@ -162,6 +162,8 @@ public:
     return asDerived().data()[idx];
   }
 
+  /// @{
+
   /*!
    * \brief Accessor, returns a reference to the given value.
    * For multidimensional arrays, indexes into the (flat) raw data.
@@ -172,7 +174,6 @@ public:
    *
    * \pre 0 <= idx < m_num_elements
    */
-  /// @{
   AXOM_HOST_DEVICE T& operator[](const IndexType idx)
   {
     assert(inBounds(idx));
@@ -354,6 +355,7 @@ public:
   /// @}
 
   /// \brief Swaps two ArrayBases
+  /// No member data, so this is a no-op
   void swap(ArrayBase&) { }
 
   /*!