+++ src/axom/quest/MeshTester.hpp
@@ -27,6 +27,11 @@
   #include "axom/mint/execution/internal/structured_exec.hpp"
 #endif
 
+// Umpire
+#if defined(AXOM_USE_UMPIRE)
+  #include "umpire/strategy/QuickPool.hpp"
+#endif
+
 // C/C++ includes
 #include <utility>
 #include <vector>
@@ -125,10 +130,9 @@ void findTriMeshIntersectionsBVH(
        : rm.getAllocator(axom::execution_space<ExecSpace>::allocatorID()));
 
   umpire::Allocator pool_allocator =
-    rm.makeAllocator<umpire::strategy::DynamicPool>(
-      allocator.getName() + "_POOL",
-      allocator,
-      POOL_SIZE);
+    rm.makeAllocator<umpire::strategy::QuickPool>(allocator.getName() + "_POOL",
+                                                  allocator,
+                                                  POOL_SIZE);
 
   const int poolID = pool_allocator.getId();